requirement.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.25.0
transformers>=4.38.0
accelerate>=0.21.0
pillow
numpy
accelerate
opencv-python
pandas
scikit-image
torchmetrics>=1.3.0
pytorch-ssim
matplotlib
jupyter
transformers
hf_xet